fix build error from Card component being in Resources folder
