+++ src/components/molecules/CardResource/styles.module.scss
