+++ src/components/molecules/index.js
@@ -10,6 +10,7 @@ import CardDefault from './CardDefault'
 import CardLexicon from './CardLexicon'
 import CardHidden from './CardHidden'
 import ColorField from './ColorField'
+import CardResource from './CardResource'
 import ColorSwatch from './ColorSwatch'
 import Date from './Date'
 import DoDont from './DoDont'
@@ -40,6 +41,7 @@ export {
 	ColorSwatch,
 	Date,
 	ColorField,
+	CardResource,
 	DoDont,
 	FormNewsletter,
 	FancyFooter,
