--- src/pages/resources/research/components/_index.js
@@ -1,3 +1,0 @@
-import Card from './Card'
-
-export { Card }
